requirements.txt
fnmatch
math
random
os
sys
time
operator
opencv-python
gc
numpy
torch
torchvision
Pillow
tensorboardX
subprocess
matplotlib
tqdm

